Adding the static apples & plate
The apples and plate were positioned on the screen, the project was re-created as there were unknown issues with building. Instead of working directly from the GitHub repository folder on my computer I am now creating the project external to the GitHub folder on my computer and then uploading changes as I go.