Buttons now react to being clicked
In MainActivity we find each button via findViewById, and activate set on click listener which listens for clicking activity. the onClick function is implemented at the bottom of MainActivity which switches cases between the different numbers and provides a different reaction dependent on the button (technically image) clicked.

With this I will now be looking to implement the answer provided in generateQuestion as a global variable so that we can reference the answer to provide different outputs (correct, incorrect).