Using the global integer to store the result and reference it!
Instead of storing the answer to a local variable 'answer' we now store it to GlobalC.ans, our newly created global variable. Now we can compare the button clicked with the answer to the question and know if it's correct or not. This is the basic functionality to next create the correct/incorrect screen.