Fixing repetitive crashing bug
When testing the application, answering an incorrect question firstly and then clicking the reset button would crash the application.

I realised this would only occur when the incorrect answer was provided firstly, so I began investigating the variables that were being recalled during resetView(); which led to investigating the clearAnimation functions for the stars.

As by default the stars were not animated the script was calling for their animations to be cleared which wasn't active, after troubleshooting I created a variable 'GlobalC.res' which is equal to 1 when an answer is correctly answered, so when resetView(); is called the star animations will only be cleared if the answer was correctly answered - afterwards setting the value of GlobalC.res back to 0.

This successfully stopped the repeated crashing.