Introduction of score keeping and game progress, made min number on random gen to 1
The score is adjusted per correction answered correctly, adding 1 to the score. When the question is answered incorrectly the progress still goes up by 1 but the score does not go up, so when the questionProg reaches 10 the score will reset to 0 and hopefully notify the user of their score (such as 7/10) before the next question pops up.

The minimum number on generateQuestion was changed to 1 as more often than not the question would generate questions such as 8+0, 6+0, etc which no longer happens here. The score is reflected after each question on the bottom left of the screen