Extended the score and questionProg code, implemented game over functionality
Basic functionality for game over has been implemented, as in when 10 questions have been answered the app will state "Game Over! Score: X/10" where X is the number of questions correctly answered (a correct answer means +1 score). An applause sound effect also plays when the game is over.

Once the game is announced as over and the user clicks the play button it will then reset the score to 0, questionProg to 0, and will essentially start as a new game.