Implemented different difficulty setting
As a basic "level" implemention I have added the global variable 'difficulty'. By default the difficulty is 0 so when the application is loaded it will be the easier setting, however if the user scores 8 or above on the first run they will be upgraded to difficulty 1 (variable changes to 1) which can provide more challenging questions for the user.

If however the user then scores less than 8 on the harder setting it will set them back to difficulty level 0 which provides easier questions (the variable is set to 0 again, so the questions will generally be easier as shown in the comments)